+++ src/application.hpp
@@ -41,7 +41,7 @@
 
 GLOBAL constexpr int EDITOR_MAJOR = 1;
 GLOBAL constexpr int EDITOR_MINOR = 0;
-GLOBAL constexpr int EDITOR_PATCH = 3;
+GLOBAL constexpr int EDITOR_PATCH = 4;
 
 FILDEF void init_application (int _argc, char** _argv);
 FILDEF void quit_application ();